+++ plugins/wqy-font/src/main/java/net/cassite/desktop/chara/plugin/wqyfont/WqyFontPlugin.java
@@ -23,7 +23,7 @@ public class WqyFontPlugin implements Plugin {
     }
 
     @Override
-    public double priority() {
+    public int priority() {
         return GENERAL_FONT_PRIORITY;
     }
